requirements.txt
djangorestframework == 3.14.0
dj-rest-auth == 5.0.2
django-allauth == 0.61.1
django-filter == 24.2
markdown == 3.0.0
drf-yasg == 1.21.7